feat: update tests after implementation of GamerService, to emulate Lottery, and update datebase from int to decimal money fields.

+++ BackEnd/TgLab.Application/Bet/DTOs/BetDTO.cs
@@ -6,9 +6,9 @@ namespace TgLab.Application.Bet.DTOs
     {
         public int Id { get; set; }
         public Currency Currency { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public BetStage Stage { get; set; }
-        public int Bounty { get; set; }
+        public decimal Bounty { get; set; }
         public DateTime Time { get; set; }
     }
 }
