+++ BackEnd/TgLab.Application/Wallet/Interfaces/IWalletService.cs
@@ -5,7 +5,7 @@ namespace TgLab.Application.Wallet.Interfaces
     public interface IWalletService
     {
         public Task Create(CreateWalletDTO dto);
-        public Task DecreaseBalance(int Id, int amount);
-        public Task IncreaseBalance(int Id, int bounty);
+        public Task DecreaseBalance(int Id, decimal amount);
+        public Task IncreaseBalance(int Id, decimal bounty);
     }
 }
